chore: fix missing unused variables from previous clippy failures
Also removes the unused `GameServerPort.name` struct field in the `agones-rest` crate that we no longer use in our code anyway to prevent further clippy warnings.

+++ crates/zohar-core/src/infra/endpoint.rs
@@ -58,7 +58,7 @@ async fn resolve_from_agones() -> anyhow::Result<SocketAddr> {
             continue;
         };
         let address = status.address.trim();
-        let Some(port) = status.ports.first().map(|entry| entry.port as u16) else {
+        let Some(port) = status.ports.first().map(|entry| entry.port) else {
             if Instant::now() >= deadline {
                 return Err(anyhow!("agones gameserver has no allocated ports"));
             }
